+++ test/utils.dart
@@ -2,12 +2,24 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+import 'dart:async';
+
 import 'package:dart2_constant/convert.dart' as convert;
 import 'package:test/test.dart';
+
+import 'package:sass/src/io.dart';
 
 /// A regular expression for matching the URL in a source map comment.
 final _sourceMapCommentRegExp =
     new RegExp(r"/\*# sourceMappingURL=(.*) \*/\s*$");
+
+/// Returns a [Future] that waits long enough for modification times to be
+/// different.
+///
+/// Windows (or at least Appveyor) seems to require a more coarse-grained time
+/// than Unixes.
+Future get tick =>
+    new Future.delayed(new Duration(milliseconds: isWindows ? 1000 : 50));
 
 /// Loads and decodes the source map embedded as a `data:` URI in [css].
 ///