+++ CHANGELOG.md
@@ -14,6 +14,14 @@
   `sass templates/stylesheets:public/css` compiles all non-partial Sass files
   in `templates/stylesheets` to CSS files in `public/css`.
 
+* Add an `--update` flag that tells Sass to compile only stylesheets that have
+  been (transitively) modified since the CSS file was generated.
+
+### Dart API
+
+* Add `Importer.modificationTime()` and `AsyncImporter.modificationTime()` which
+  report the last time a stylesheet was modified.
+
 ### Node API
 
 * Generate source maps when the `sourceMaps` option is set to a string and the